tools/tkl-float/requirements.txt
PySide6==6.6.0
requests==2.31.0
python-dotenv==1.0.0
keyboard==0.13.5
pywin32==306; sys_platform == "win32"
